requirements.txt
requests
openai==1.3.9
duckduckgo-search==0.9
beautifulsoup4
python-dotenv
livekit-agents==1.1.5
langchain